+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+
 ### Fixed
 - Evaluating freshly initialised encoder models on multiple-choice classification tasks
   caused an error, as the id-to-label mapping was not set up correctly. This has been
@@ -15,6 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added the English knowledge dataset Life in the UK. This was contributed by
   [@oliverkinch](https://github.com/oliverkinch) ✨
+- Added the Norwegian idioms dataset, which is a multiple-choice question dataset,
+  where the alternative answers have been generated using GPT-4o.
+  This was contributed by [@oliverkinch](https://github.com/oliverkinch) ✨
+
 
 ## [v15.10.1] - 2025-06-20
 ### Fixed